requirements.txt
flask==3.1.3
tensorflow-cpu==2.15.0
numpy==1.26.4
opencv-python-headless==4.8.0.76
scikit-learn==1.3.2
matplotlib==3.7.5
seaborn==0.13.2
Pillow==10.4.0
gunicorn==21.2.0
keras==2.15.0
scipy==1.11.4
gdown==6.0.0
